starters/od-project-starter/od-project/requirements.txt
ultralytics>=8.2.0
torch
matplotlib
pyyaml
numpy
